scripts/requirements.txt
schematics~=2.1.1
pyyaml~=6.0.2
fuzzywuzzy~=0.18.0
pluralizer~=1.2.0
lxml>=5.0.0
flask~=3.0.3
cachelib~=0.13.0
xmltodict~=0.13.0
packaging>=21.3
Jinja2~=3.1.4
MarkupSafe>=2.1.5
jsonschema~=4.23.0
click~=8.1.2
setuptools==70.0.0
python-Levenshtein
azure-mgmt-core
wrapt>=1.11
